Open/Close methods from QSerialBusDevice now opens/closes backend connection
SerialBusBackend interface class now has open(QIODevice::OpenMode) and close()
methods. When open() in QSerialbusDevice is called, open in plugin is called
in addition to QIODevice open(). This gives the plugin developer possibility
to make bus connection at this point. Same functionality works with close().

In SocketCanBackend connection to socket is now done in open() instead of
constructor. close() closes the socket.

Change-Id: Iba8858bbe78c419181181eac281bd9a8c426b378
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>