Remove the datastream versions
We are using the data stream version of the Qt used at runtime. The plugin
and the library always read and write to data streams using the same
Qt version. Stream versions only make sense when saving the stream to a
file and reading it with another application (using a different
Qt version).

Change-Id: I31eed483f1b58bbbead03a2dc751562455fcfe10
Reviewed-by: Janne Heikki Partala <janne.partala@bitfactor.fi>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>