Separate frameId from ETR/EFF/ERR bits
This is a step towards removing can.h/Linux specific elements from
the API. We are adopting though the Linux specific bit format for
the frame identifier.

The frameId becomes a uint. A placeholder and versioning was added
for later member additions.

Since the frame format is changing this patch invalidates all existing
streaming code for QCanFrame. This patch converts all
used/four streaming operators against the official ones in QCanFrame.

Change-Id: I5f82ce16091712933b4a7a0edb61145144b162c0
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>