Adjust config param handling in socketcan
The configuration parameter can be applied before and after a connect.
So far the example application always handled the parameters after
a connect. The recent can example changes set the params before connect.
This lead to the result that the new params were ignored.

This change makes it possible to store the params until they are needed.
In the case of socketcan the params can only set after the connect
when the can bus socket becomes valid.

Change-Id: I983fa7d41a63d8f806d0510b4c013d967f039d51
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>