Add support of Linux API for the PeakCan plugin
The PCAN uses same API (the functions and the library names) for any of
supported platforms (Windows and Linux). A small differences is only in
the read event handling. So, it is make sense to remove the *_win and
the *_stub files, and to move all private code to one cpp file.

Change-Id: I252c85ae15f69d53997ce87dfb528bbacb71bc9c
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>