Optimize the CAN frames receiving
We do not need to emit a signal when each frame is received,
this increases the load of the event loop. The input FIFO can
contains more than one frame, thus we only need to emit the
signal when all current frames were read.

In this case this signal is renamed to framesReceived instead
of frameReceived.

Change-Id: If97cfd4304dae9549748f87dbd0b54eae4c0a0f2
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>