Initial import of libmodbus version 3.1.2
Link libmodbus statically against the related modbus plugin.
This enables us to deploy the library on all platforms.
Note: In order to do so a small patch had to be applied for
Windows. See the patches folder while upgrading the library.

The library is LGPL 2.1 or later. Therefore we cannot statically
link it against the library. The libmodbus plugin
has to remain forever. However the plugin will eventually
be invisible to the user.

Another commit will add the option to dynamically link
against the platforms version.

Change-Id: I1e08ea90f8517afa38e059aed35441dc048286ff
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>