Add NegativeAcknowledge (7) missing exception code
Add NegativeAcknowledge to the enum of known exception codes. The code
is not included in the current modbus spec but is widely used in devices
that use it as a response for programming commands (such as Schneider
Electric). The according negative acknowledge message counter is also
already present which accompanies this exception code.

Change-Id: Ida86914bfe073601e2a286d92978f547e3ae7b60
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>