Fix documentation and add info on createReadWriteRequest
Fix documentation where the slaveId was still used instead of
slaveAddress and add information about the used function code and the
according behavior. Modbus specifies that write is operated before read,
but devices may behave differently so the results on overlapping data
areas for the read/write operation needs to be taken into account.

Change-Id: Iec79e15c57a9ec2eeaa003e880d912a223e8aaef
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>