Improve the QModbusPdu and derived classes documentation
Variadic templates cause qdoc to choke. Documenting them at all is
challenging at best. The variadic ctor's will not be documented
whereas the two public function (using variadic templates) will be
documented. The ctors have been shifted because qdoc stops working
once it encounters such a template. This would cause subsequent
functions in the same class to not appear at all.

The signature of the to-be-documented functions has been simplified
so that qdoc recognizes them. It causes some information loss but the
attached code snippets should cover most issues.

The not documented ctors remain in the code base to keep the warnings
in the qdoc output.

Every other qdoc warning related to the same classes has been fixed.

Task-number: QTBUG-47085
Change-Id: Iff5355fa62d6a54feec565f52c5869c973cb1514
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>