Do not register the generic plugin for automatic linking.
Issuing a distclean after a static build (tests + examples),
the build system will not cleanup all generated .pri files
from qtbase/mkspecs/modules (known issue).
On a second build (-tests) the qt_plugin_qtcanbustestgeneric.pri
will be found and included as plugin to link against, though now
it is not build as it is located in tests.

Change-Id: If9d6b3f7d8d7a38af01c6792b276099537da9340
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>