CAN-Example: Add hint to ConnectDialog interface edit
It was really hard to figure out what to put in the interface
edit without deeper knowledge of the backends.

Now the interface edit a placeholder with the first possible
name every time a new backend is selected.

This is thought as a temporary fix, the proper solution
would be to get the possible interfaces from the backends
itself.

Change-Id: I74b5c072d64f7ee98d3fa65e059e6b00c510b308
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>