CAN: Fix displaying extended identifiers below 0x800
While a standard identifier has a range from 0 to 2^11,
extended identifiers have a range from 0 to 2^29.

As they are distinct by the IDE (identifier extension) bit,
the same numeric value is still different in the overlapping
range from 0...0x7FF.

E.g. 0x123 (IDE = 0) is different from 0x00000123 (IDE = 1).

With this change, the received messages are displayed
correctly according to the IDE bit.

The proper solution is already done in dev branch as
55ad040de1ac6a33b4f2efe29038e97e8d43dbc2

Task-number: QTBUG-53171
Change-Id: I7383b0f9c5e05746f442998cddfafe32f0932120
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>