CAN: Prevent to open or close an already opened or closed device
We do not need do excess checks on is "open/close", as this
should becomes once in QCanBusDevice.

* QCBD::connectDevice() should return an error when the device
still is in progress of connecting/disconnecting, or already
connected.

* QCBD::disconnectDevice() has to make disconnection when the
device is in connected or connecting state; otherwise if the
device is in disconnecting state, then it should do nothing
(and as well when the device is in disconnected state).

Note: This behavior is similar to QAbstractSocket.

Change-Id: I4b62e728d37b190fbaee768bdd0a2878c3ba98ab
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>