CAN: Unify interface names
Mostly, the interfaces now simply start with can0 and the number
increases for every next interface.

Some vendors distinguish between e.g. PCI and USB interfaces. In this
case, the interfaces are named usb0 and pci0, with increasing numbers,
respectively.

Multiple channels per device are accessed by adding the channel number
after a dot: can0.0 for the first channel in the first device (TinyCAN).

SocketCAN is still a bit different, as it allows arbitrary interface
names to be mapped to the hardware or virtual CAN busses.

[ChangeLog][QtSerialBus][Unify interface names] The interface names
of the different backends were unified to follow a common rule: The
prefixes were unified, the numbering now always starts with zero.

Change-Id: Ia2ca76a9be01b9123aed1ede90debf69871bb82c
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>