CAN: Add QCanBusDeviceInfo
Returns:
* a list of available devices for the chosen plugin
* if the device is virtual
* if the device is CAN FD capable

Use this list in the CAN Example for interface selection.

If information about virtual channels or CAN FD capabability is not
available, false is returned.

For SocketCAN, PeakCAN, VectorCAN and SystecCAN, the implementation is
completed. TinyCAN only return the list of possible interfaces for now.
This is can be changed later, but needs quite some refactoring of the
TinyCAN plugin.

[ChangeLog][QtCanBus] Added the new class QCanBusDeviceInfo
for enumeration of available CAN interfaces and more information
about them.

Task-number: QTBUG-54298
Change-Id: I851bcc3b9ee41aaaf1164c6b4a5aaf6503cd8746
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>