QCanBusFrame: Introduce local echo flag
Inspiration for change after reading the SocketCAN mailing list thread:
https://marc.info/?l=linux-can&m=147438437202356&w=2

The main usage is, to verify at application stage, that and when a
frame was really sent on the bus. For architectures like SocketCAN,
where multiple applications can connect to one CAN bus on the same
system, this flag also indicates frames that are sent by an other
local running application (i.e. not coming from the real bus).

Implementations done for SystecCAN, VectorCAN and SocketCAN plugin.
PCAN Basic does not seem to have support for this flag.

For SocketCAN, see section "4.1.7 RAW socket returned message flags"
in https://www.kernel.org/doc/Documentation/networking/can.txt
for reference.

[ChangeLog][QCanBusFrame] Added local echo flag to mark all echo frames
that are generated by the hardware on successful transmission to CAN bus
as such.

Change-Id: I66d4ce1ec6dfd0a0b994a7b7a9de1a5f503fcb70
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>