CAN-Example: Add a dedicated send frame group box
By moving the send frame logic to a separate panel, the MainWindow
gets easier to understand as it only connects signals to slots.

The new SendFramesBox also does more sanity check on the parameters
(options that exclude each other are checked) and forbids invalid input
by using QValidators. The payload input is also prettified with spaces.
This was first done by InputMasks but this had strange side effects
so now the input validator is used for this task also.

Change-Id: Ib95605ce4bc23b52d1cf5d082b1533b2a59f7749
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>