Fix unxepected state in Modbus process queued element
Due to the nature of our serial bus device, a frame could already
be send completely, but the bytes written did not report that yet.
Then we received an answer before we reached the receive state,
messing up the whole state logic. Now we process incoming frames
only if we are in receive state, otherwise drop full frames.

Task-number: QTBUG-62144
Task-number: QTBUG-62299
Change-Id: I88a4cf0b64a823409cbb277431a004a9d126cddc
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>