PeakCAN: Always compile without PCANBasic.h
Keep the possibility to directly link against pcanbasic.dll/so for now,
as it allows to use the PeakCAN plugin without requiring QLibrary.

However, as we need to declare all symbols in peakcan_symbols_p.h anyway,
there is no need to require the header file.

And I need to add another symbol in a follow-up to resolve Clang warnings.

Change-Id: Idb7e57d40c8074c0d14d0d1525b09663d148cb8b
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>