Add J2534 Pass-Thru CAN plugin
[ChangeLog] Add J2534 Pass-Thru CAN plugin

SAE J2534 is a standard for communication between a computer and
a vehicle. The "passthrucan" plugin adds support for communicating
with a vehicle's CAN bus via the generic J2534 Pass-Thru API.

The J2534 Pass-Thru API v04.04 is only specified for 32-bit Windows,
but implementations for 64-bit Windows and other platforms do exist.
Automatic device discovery is based on the Windows registry and thus
supported on Windows only. However, this plugin alternatively allows
users to specify the path to a J2534 vendor library to load.

Any CAN interface adapters that ship with a compliant J2534 Pass-Thru
interface library (v04.04) for the target platform should work with
this plugin. It should be noted though that driver support for x64 is
as of yet rare, thus a 32-bit Qt build may be required to make use of
the passthrucan plugin.

Change-Id: I55280d455828d9cf0ad03690224b924b39c17cf7
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>