Fix RTU Master ignoring responses due to being in Send state
The fix for QTBUG-62299 limits the processing of serial responses
in RTU serial master to cases when the state machine is in receive mode
only. So far the Receive state was triggered by the sendTimer. However
modern kernels accept serial packages faster then the sendTimer can fire.

This patches addresses the problem by setting the Receive state even before
sendTimer fires. If the bytesWritten() callback detects that the entire
PDU/ADU was written it stops sendTimer and immediately progresses the
state machine to the Receive state.

Fixes: QTBUG-69188
Change-Id: I1657973ef99f1b3e4008e887029b30890d825d59
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>