Permit dropping of RTU slave request buffer under certain scenarios
If there is a wrong character in between individual modbus PDUs the
request buffer becomes longer and longer w/o ever being cleared.
As a consequence the only possible way to recover is to reset the RTU slave.

The patch ensures that request buffer processing waits for chunked PDUs
(due to low baud rates) while at the same time does not block if it
contains garbage. The timer is based on the modbus spec which specifies that
every modbus package is separated by at least 3.5 chars. Using the current baud
rate of the serial port permits the exact caluclation of the implied
delay/timeout.

For baud rates above 19200 the timeout drops below 2 ms which is not
measurable by Qt anymore. Furthermore the modbus spec specifies a minimal
delay of 1.75 ms as well. Hence the timeout is bound to at least 2ms.

Fixes: QTBUG-59286
Change-Id: Ie43ff3f5819519abe2081757150d367820d50020
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>