Fix case of accessing deleted memory
If the current QModbusRtuSerialMaster instance is deleted, the
interFrameDelay timer may still be active and trigger the associated
lambda. Since the lambda calls back into the deleted class instance,
a crash may occur. Adding the public class instance as context to
the singleshot connect ensures that the lambda is not called if the class
instance was deleted already.

Fixes: QTBUG-65684
Change-Id: Ic7d90ba68a1533e6de50c61117f5de8dc5e048b4
Reviewed-by: André Hartmann <aha_1980@gmx.de>