Systec: Fix error message "Cannot configure TxEcho for open device"
When using the QCanBusDevice::ReceiveOwnKey, it has to be applied
before the device is opened. While this already correctly happened
in SystecCanBackendPrivate::open(), the parameter was applied again
in SystecCanBackendPrivate::setConfigurationParameter() which led
to the error message.

Solution is to skip ReceiveOwnKey when applying all configuration
parameters in SystecCanBackend::open(), like already done there
for BitRateKey.

[ChangeLog][SystecCAN] Fixed the error message  "Cannot configure
TxEcho for open device" when QCanBusDevice::ReceiveOwnKey was used.

Change-Id: I37c1732798a3ce205a65ca803a99b5b7b9306d2a
Reviewed-by: Samuel Gaist <samuel.gaist@idiap.ch>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>