QCanBusDevice: Make Filter comparison functions non-member friends
Symmetric operators, such as op== and op!=, shouldn't be members,
because member functions do not allow for implicit conversions on
the LHS argument, which introduces an asymmetry in how the RHS and
LHS arguments are handled by the compiler.

We could add the operators as non-member non-friend free functions,
but since Filter is a nested class, there's a lot of code between
the definiton of Filter and the point where we can lexically declare
the operators for the first time (after the definition of the outer
class). Code between these points does not see the operators and thus
behaves different from code that follows the the definition of the
outer class. This is subtle and while in the present case, there's
no indication that Filter may get an implicit conversion to some other
type which does have an equality operator, there's always the
possibility that such a thing might be added later, in which case code
using lhsfilter == rhsfilter in the body of QCanBusDevice would call a
different operator== than code outside it.

It's just safer to declare the operators as soon as possible, which
means as non-member friends.

Change-Id: Iadd097a9d64d0b1c912b750d6d89431b42def313
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>