VirtualCAN: Fix disconnecting from bus
While the client sent a disconnect command to
the server, this was not handled on server side.
Thus, the client was never really Unconnected and
in turn it was not possible to connect again.

Fix that by closing the socket, that leads to
VirtualCanBackend::clientDisconnected() being
called and that sets the plugin to Unconnected.

In any case, the server keeps running until the
plugin is unloaded, even after the last client
is disconnected.

[ChangeLog][VirtualCAN] Fixed that the plugin
stayed forever in closing state and subsequent
connect calls always failed.

Change-Id: I7a490c55a3cb6bd193ba4783b48d1385aa85bcde
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Rolf Eike Beer <eb@emlix.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit f47b712057676de6115d7320cbbc2eb9b20c9817)