CAN-Example: Fix two issues with the status timer
1. Fix a crash when device open fails

If opening of m_canDevice failed, m_canDevice is reset to
nullptr in line 154. Dereferencing that leads to a crash.

2. Avoid connecting multiple slots to status timer

Opening and closing multiple times lead to multiple
connections of the lambda slot, which triggered the
call to m_canDevice->busStatus() unnecessarily often.

The solution is to do the connection only once,
and stop the timer when m_canDevice becomes nullptr.
It is started again after a successful open.

Amends f83587a9978bd212d

[ChangeLog][Examples] Fixed a crash in CAN-Example when
opening the CAN bus device failed.

Change-Id: I4d19f382593fa26fa3b8166f7121e7ff14fd2e74
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>