Generic(V1): Fix emitting framesReceived when disconnected
When one of these plugins was used in the
CAN-Example, the frames received slot was
still called after disconnecting.

That made the example stuck and emitting
lots of "Cannot read frame as device is
not connected." messages.

This is mostly relevant for developers
though, as these plugins are not deployed.

Change-Id: I675518148f4bf37f11911dd7a0f6423ae18be19d
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>