QCanBusDevice: Remove outdated comments from public header
While (Dis)connectDevice is chatty for a class already
containing the word device, we cannot simplify that as
QCanBusDevice inherits from QObject, so connect/disconnect
are already allocated. So let's just keep these as-is.

Also, the open and close calls should just be kept and called
as up to now.

Change-Id: I85febbd2253753a67058ae8ff097635106853e77
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>