Update commercial license headers
Updated header.COMM to the files in tqtc-qtserialbus. Examples,
tests and documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals
with the earlier license header. Reason for this is
is that some autotests use hard coded line numbers
and a change in the line count causes failures in tests.

Task-number: QTQAINFRA-4216
Change-Id: Ib339fd35b4c59daa82458bd8796b967f2dfee1ca
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>