Fix Modbus custom command size calculation
We cannot early return with an invalid PDU if the private
minimumDataSize() functions returns a negative value. The
negative value might mean we hit a user defined function code,
so advance and call the user data size calculater if registered.

The broken function was only used by Modbus TCP implementations ,
so RTU was working correctly since it created the PDU differently.

[ChangeLog] Fix Modbus TCP custom command size calculation.

Pick-to: 5.15
Fixes: QTBUG-62192
Fixes: QTBUG-91037
Task-number: QTBUG-91214
Change-Id: I2d421e073777306d86a83cd05a289481fa082501
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>