Introduce QCanBusFrame::FrameId
Use a dedicated type instead of quint32
to mark a CAN bus frame identifier as such.

For now it is just an alias to quint32, but
can be changed to some other type for Qt 7.

[ChangeLog][QCanBusFrame]
Introduced a specific type QCanBusFrame::FrameId
for the CAN bus frames identifier. The new type
is a typedef to quint32 and therefore should be
source compatible.

Change-Id: Ie26aa69cd48eb9cee4c11a0c8f8ba74efae24b4f
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>