Add test case to show different behavior in read functionality
All read functions end up in either collateBits() or collateBytes().
The latter seems to ignore all values given by the Modbus data unit
and fills it based on the information received by the Modbus response.
collateBits() on the other hand ignores all information in the response
and tries to fill the data unit based on the values in there.

Task-number: QTBUG-80542
Change-Id: I60dfdc83ab1be9c832e834982376fee4c3e5ba7c
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>