CMake: Explicitly link against the built QtNetwork library
This is to ensure that the pre-existing QtNetwork library from the
b2qt SDK sysroot in the CI is not picked up.

This is likely a consequence of CMake not adding an -rpath-link flag
when linking. See QTBUG-86533 for details.

Task-number: QTBUG-86533
Change-Id: Id05d49812f957c8a6c12c03f5ff633015bd63103
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 4a2828a3d30fa1e4760a6778b8b58569caefb632)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>