Update commercial license headers
Updated header.COMM to the files in tqtc-qtserialbus.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4960
Change-Id: Id8e2889a6a9bcf9bf11edc479e72b80c3e232676
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
Reviewed-by: Antti Kokko <antti.kokko@qt.io>