Turn canbusutil into app
A tool is normally not cross-compiled, but canbusutil is meant to be
used in the target env. Turn it into app so that it's always compiled,
except for iOS, WASM and rtems, where it's not expected to work.

Change-Id: I4286ea524a05727c237ec68454e5406f96d7a7aa
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>
(cherry picked from commit c84d82c8ae24e59c181aea48b8884ac60c7484a7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>