QtSerialBus: replace qSwap with std::member-swap where possible
qSwap() is a monster that looks for ADL overloads of swap() and also
detects the noexcept of the wrapped swap() function, so it should only
be used when the argument type is unknown. In the vast majority of
cases, the type is known to be efficiently std::swap()able or to have
a member-swap. Call either of these.

Task-number: QTBUG-97601
Pick-to: 6.3 6.2 5.15
Change-Id: I89336e147d12e6d6563fab2e76f5ae7eec4ff220
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>