Long live QCanFrameProcessor
[ChangeLog][CAN Bus] Introduce the QCanFrameProcessor class.
This class can be used to decode the received QCanBusFrame into
a map of key-value pairs, and also compose a QCanBusFrame out of
a map of key-value pairs.
The class uses the information provided by QCanUniqueIdDescription
object and a list of QCanMessageDescription objects as rules to do
the encoding or decoding.

Task-number: QTBUG-98910
Change-Id: I9a6fc9b681ef4f0a47bc659a0536277c03379d09
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>