QCanFrameProcessor: update BE data handling
This commit changes the way of encoding and decoding the big endian
data. It also updates the documentation for the QCanSignalDescription
and QCanUniqueIdDescription classes, explaining how to properly
define parameters for the LE and BE cases.

This commit also refactors the unit tests. Previously we had separate
tests for checking encoding/decoding into Payload and FrameId. Now
they are joined, and an extra parameter (source) is introduced. This
allows to avoid some data duplication. Also adds the roundtrip test.

Note that the old was of data processing is not completely removed,
but hidden under the ifdef check. The reason for that is that we
want to get some feedback and see, if we need to support both ways
at the same time.

This commit amends fb5b117312df3e156beac0aebbd9891571dfa8bd.

Task-number: QTBUG-98910
Change-Id: I5c0d2e561b92b409a8e15c27bd4e65ad7dba3081
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>