Simplify QCanFrameProcesor
This class is not a value class. Normally the user will need to create
only one instance of this class and use it to process all messages.
Because of that, we do not really need to support copy and move operations.

This allows us to significantly simplify the class:
* remove copy/move constructors and assignment operators
* remove {in}equality operators
* use std::unique_ptr instead of QESDP for d_ptr
* use Q_DISABLE_COPY_MOVE

Change-Id: Idf9844f783700a82d2740858c76abed1696d5b3c
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>
(cherry picked from commit 4d037d7faef8266c79e25dfeaf5f9ecd9b6d1bc1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>