QModbusTcpServer: fix memory leak
Noticed while improving Modbus Custom Command example.

Before the patch the QTcpSocket::destroyed signal was connected to
a lambda in context of the parent QTcpServer (the 'q' variable).

When the application was closed without disconnecting the server in
advance, the QTcpServer::close() was called in the server's destructor.
The client sockets were disconnected, but by the time their destructors
were called, the parent QTcpSocket was already in its desctuctor, so
the connection was destroyed (because deleteChildren() is called after
invalidating connections in QObject's d-tor).

This patch simply changes the context of the lambda's connection to
the socket object itself, so that the lamdba is correctly invoked
from the socket's destructor.

Change-Id: I04e196645db35c390a2bb255f1fadaac336fec25
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit bfb62b60ad6873ad1912df0f08ce9cf7fd80447b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>