Doc: Use the \examplecategory macro to tag examples
The '\meta category' command was used for tagging examples with a
specific category, used in Qt Creators Welcome mode.

As we want to also generate lists of examples belonging to a category
elsewhere in the documentation, replace the command with a macro that
expands to the original \meta command and also adds the example to
a group using the \ingroup command. This way, the category names can
be used as arguments to the \generatelist or \annotatedlist commands.

Task-number: QTBUG-112731
Change-Id: Ib1b2ea3b93e0a24d365a822965fe512851568a6a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 0c45f0423d573687917f2093e7c1d87a4bf17aa7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>