QCanDbcFileParser: add parseData(QStringView)
[ChangeLog][CAN Bus][QCanDbcFileParser] Added method
parseData(QStringView), which takes the contents of a DBC file as
an input parameter. Use this method if the DBC file has an encoding
different from UTF-8. In such cases the file contents should be first
properly converted to QStringView, and then passed to this method.

Update most of the parser unit-tests to cover two cases:
* parsing a DBC file using QCanDbcFileParser::parse().
* parsing the contents of the file using new
  QCanDbcFileParser::parseData().

Fixes: QTBUG-115281
Change-Id: Ie9743c29e64ebb36ce57372cf23c6e1288e98b9e
Reviewed-by: André Hartmann <aha_1980@gmx.de>