VectorCAN: Set CAN-FD configuration only for uninitialized channels
The Vector plugin allows multi application access to single CAN
channels, but only the first application gets init access and
is allowed to set the configuration parameters.
On hardware devices you get an invalid access error, but the
virtual channels return an XL_ERROR and the connection fails.

[ChangeLog][CAN][Vector] Fixed XL_ERROR when try to set CAN-FD
configuration for a virtual channel which was already opened by
another application.

Fixes: QTBUG-123012
Pick-to: 6.7
Change-Id: Iafb770940025b5c71fd5db83e61edbb0f49254e2
Reviewed-by: André Hartmann <aha_1980@gmx.de>