Fix virtual can backend usage in multithreaded environment
Test for the current thread in VirtualCanServer::start() to prevent
that the server is started from within a different thread than the
servers one.
This situation can occur if the start() method is invoked
twice, but from different threads, and the first time the server
was created but couldn't be started successfully (e.g. because the
TCP port was already blocked by a virtual CAN server instance in
another process).

[ChangeLog][CAN Bus][Virtual CAN] Fixed a race condition
during start of the VirtualCanServer in multithreaded
environment

Pick-to: 6.8 6.7
Change-Id: Ia0f80e491accac3151d40d6978fbdd91f38a97d6
Reviewed-by: André Hartmann <aha_1980@gmx.de>