VirtualCAN: Add support for 10 channels
There is no technical limitation for the channel number,
but there are usages where two channels were not enough.

Extend it to 10 for now as this avoids modifications of the
regular expression code `\d`, but in practice it can be any
larger number also.

[ChangeLog][VirtualCAN] The VirtualCAN plugin now supports
up to ten virtual CAN channels from can0 to can9.

Change-Id: I78d6f7cd7108ae82493368344bb4c1dcd0504b4d
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>