Mark CAN parsers as security-critical
The qcandbcfileparser.cpp file contains the parser for the user-provided
DBC files.
The qcanframeprocessor.cpp file parses the CAN payload to extract the
values based on the message descriptions. The parsing involves some
low-level buffer and bit-fiddling, and the data can come from the CAN
bus, or be provided by the users.

This patch marks both of the files as security-critical according to
QUIP-23.

Task-number: QTBUG-135465
Pick-to: 6.9 6.8
Change-Id: I162afca20029919ec061dc9f2be6ce9275dfa5dd
Reviewed-by: André Hartmann <aha_1980@gmx.de>