Mark qmodbus{client,server}.cpp as security-critical
Both files are parsing the raw data from the QModbusPdu struct and
extracting values from it. The struct can be populated with the data
received from the other devices, as well as manually constructed by the
users, so we need to make sure that we parse the data correctly.

Task-number: QTBUG-135465
Pick-to: 6.9 6.8
Change-Id: If6973f65af20a6ce66cc64b23a0afac3bae62b7f
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>