Mark Modbus RTU serial client and server as security-critical
The qmodbusrtuserial{server,client}_p.h headers contain the code that
receives raw data from a serial port and tries to parse it in order to
create a QModbusPdu object.

The qmodbusrtuserialserver.cpp file contains a custom implementation
of processRequest(), that parses raw data from the QModbusPdu struct.

Task-number: QTBUG-135465
Pick-to: 6.9 6.8
Change-Id: I0a8c31f7dbc3c8c8cf9fc6b70286361a2503f539
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>