Mark Modbus TCP client and server as security-critical
The qmodbustcp{server,client}_p.h headers contain the code that
receives raw data from a tcp socket and tries to parse it in order to
create a QModbusPdu object.

The qmodbustcpserver.cpp file contains a custom implementation
of processRequest(), that parses raw data from the QModbusPdu struct.

Task-number: QTBUG-135465
Pick-to: 6.9 6.8
Change-Id: Iae95beae2577d4e43369dbb97c9a69d0f46f4832
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>